+++ pwndbg/aglib/typeinfo.py
@@ -7,9 +7,9 @@ from __future__ import annotations
 import sys
 from typing import Dict
 from typing import Optional
-import pwndbg.dbg_mod
 
 import pwndbg
+import pwndbg.dbg_mod
 
 module = sys.modules[__name__]
