+++ pwndbg/auxv.py
@@ -9,7 +9,9 @@ from __future__ import annotations
 
 import re
 import struct
+from typing import Dict
 from typing import Optional
+from typing import Union
 
 import pwndbg.aglib
 import pwndbg.aglib.file
@@ -18,10 +20,11 @@ import pwndbg.aglib.qemu
 import pwndbg.aglib.stack
 import pwndbg.aglib.typeinfo
 import pwndbg.color.message as M
+import pwndbg.dbg_mod
 import pwndbg.lib.cache
 import pwndbg.lib.config
 from pwndbg.lib.elftypes import AT_CONSTANT_NAMES
-from pwndbg.lib.elftypes import AUXV
+from pwndbg.lib.elftypes import AT_CONSTANTS
 
 # We use `info.auxv()` when available.
 if pwndbg.dbg.is_gdblib_available():
@@ -58,6 +61,43 @@ example_info_auxv_linux = """
 15   AT_PLATFORM          String identifying platform    0x7fffffffdb49 "x86_64"
 0    AT_NULL              End of vector                  0x0
 """
+
+
+class AUXV(Dict[str, Union[int, str]]):
+    AT_PHDR: Optional[int]
+    AT_BASE: Optional[int]
+    AT_PLATFORM: Optional[str]
+    AT_BASE_PLATFORM: Optional[str]
+    AT_ENTRY: Optional[int]
+    AT_RANDOM: Optional[int]
+    AT_EXECFN: Optional[str]
+    AT_SYSINFO: Optional[int]
+    AT_SYSINFO_EHDR: Optional[int]
+
+    def set(self, const: int, value: int) -> None:
+        name = AT_CONSTANTS.get(const, "AT_UNKNOWN%i" % const)
+
+        if name in ["AT_EXECFN", "AT_PLATFORM", "AT_BASE_PLATFORM"]:
+            try:
+                value = (
+                    pwndbg.dbg.selected_inferior()
+                    .create_value(value)
+                    .cast(pwndbg.aglib.typeinfo.pchar)
+                    .string()
+                )
+            except Exception:
+                value = "couldnt read AUXV!"
+
+        self[name] = value
+
+    def __getattr__(self, attr: str) -> Optional[Union[int, str]]:
+        if attr in AT_CONSTANT_NAMES:
+            return self.get(attr)
+
+        raise AttributeError("%r object has no attribute %r" % (self.__class__.__name__, attr))
+
+    def __str__(self) -> str:
+        return str({k: v for k, v in self.items() if v is not None})
 
 
 @pwndbg.lib.cache.cache_until("objfile", "start")
