+++ pwndbg/lib/elftypes.py
@@ -337,40 +337,3 @@ class Elf64_Phdr(pwndbg.lib.ctypes.Structure):
         ("p_memsz", Elf64_Xword),
         ("p_align", Elf64_Xword),
     ]
-
-
-class AUXV(Dict[str, Union[int, str]]):
-    AT_PHDR: Optional[int]
-    AT_BASE: Optional[int]
-    AT_PLATFORM: Optional[str]
-    AT_BASE_PLATFORM: Optional[str]
-    AT_ENTRY: Optional[int]
-    AT_RANDOM: Optional[int]
-    AT_EXECFN: Optional[str]
-    AT_SYSINFO: Optional[int]
-    AT_SYSINFO_EHDR: Optional[int]
-
-    def set(self, const: int, value: int) -> None:
-        name = AT_CONSTANTS.get(const, "AT_UNKNOWN%i" % const)
-
-        if name in ["AT_EXECFN", "AT_PLATFORM", "AT_BASE_PLATFORM"]:
-            try:
-                value = (
-                    pwndbg.dbg.selected_inferior()
-                    .create_value(value)
-                    .cast(pwndbg.aglib.typeinfo.pchar)
-                    .string()
-                )
-            except Exception:
-                value = "couldnt read AUXV!"
-
-        self[name] = value
-
-    def __getattr__(self, attr: str) -> Optional[Union[int, str]]:
-        if attr in AT_CONSTANT_NAMES:
-            return self.get(attr)
-
-        raise AttributeError("%r object has no attribute %r" % (self.__class__.__name__, attr))
-
-    def __str__(self) -> str:
-        return str({k: v for k, v in self.items() if v is not None})
